+++ src/completions/properties.ts
@@ -50,33 +50,37 @@ export class PropertyCompletionProvider
 		);
 
 		if (isKeyMatch) {
-			// find first unclosed bracket
-			const lastBracketIndex = combinedLines.lastIndexOf("}");
-			let bracketIndex =
-				lastBracketIndex !== -1 ? lastBracketIndex : combinedLines.length;
-			let bracketCount = 1;
-			let classNameMatch;
+			/*
+            We need to handle cases like this:
+            new("x"){
+                {
+                    // it should NOT autocomplete in here
+                }
+                // it SHOULD autocomplete in here
+            }
+            */
 
-			while (bracketCount > 0 && combinedLines[bracketIndex - 1]) {
-				bracketIndex--;
-				const char = combinedLines[bracketIndex];
+			let bracketCount = 0;
+			let classNameMatch;
+			for (let i = combinedLines.length; i > 0; i--) {
+				const char = combinedLines[i];
 				if (char === "{") {
-					// try matching with classnamematch
+					bracketCount++;
+
+					// try to match classname
 					classNameMatch = combinedLines
-						.substring(0, bracketIndex)
+						.substring(0, i + 1)
 						.match(regex.className);
 
 					if (classNameMatch) {
 						break;
 					}
-
-					bracketCount++;
 				} else if (char === "}") {
 					bracketCount--;
 				}
 			}
 
-			if (classNameMatch) {
+			if (classNameMatch && bracketCount === 1) {
 				const aliases: string[] = (
 					(vscode.workspace
 						.getConfiguration("fusionautocomplete")
@@ -141,11 +145,18 @@ export class PropertyCompletionProvider
 			if (hasValue) {
 				item.insertText = new vscode.SnippetString(property.Name);
 			} else {
-				item.insertText = new vscode.SnippetString(
-					`${property.Name} = ${
-						autocompleteText[property.ValueType.Name] ?? ""
-					}`
-				);
+				let insertText = `${property.Name} = `;
+
+				if (autocompleteText[property.ValueType.Name]) {
+					const postLineCharacters = vscode.workspace
+						.getConfiguration("fusionautocomplete")
+						.get("postLineCharacters");
+
+					insertText +=
+						autocompleteText[property.ValueType.Name] + postLineCharacters;
+				}
+
+				item.insertText = new vscode.SnippetString(insertText);
 			}
 
 			return item;